+++ include/nstd/os.h
@@ -1,4 +1,5 @@
 #ifndef NSTD_OS_H_INCLUDED
 #define NSTD_OS_H_INCLUDED
+#include "os/linux.h"
 #include "os/windows.h"
 #endif