+++ include/nstd/os/linux.h
@@ -1,0 +1,4 @@
+#ifndef NSTD_OS_LINUX_H_INCLUDED
+#define NSTD_OS_LINUX_H_INCLUDED
+#include "linux/alloc.h"
+#endif