+++ src/os/linux.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "nstd_os_alloc")]
+pub mod alloc;